Updated UI to show multiple shares Verified correctness of parser for multiple shares Parsing strike price

+++ Shares.Model/ShareDay.cs
@@ -21,7 +21,7 @@ namespace Shares.Model
         {
             var fields = new[]
             {
-                Tf.ToString("X"),
+                ((byte)Tf).ToString("X"),
                 Date.ToString("yyyyMMdd"),
                 Open.ToString("00.000"),
                 High.ToString("00.000"),
